Added regester

+++ base/views.py
@@ -5,8 +5,11 @@ from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
-
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
+
 # rooms = [
 #     {'id': 1, 'name': 'Lets learn python!'},
 #     {'id': 2, 'name': 'Design with me'},
@@ -14,8 +17,12 @@ from django.contrib import messages
 # ]
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -31,8 +38,25 @@ def loginPage(request):
         else:
             messages.error(request, "Wrong Username or Password")
 
-    context = {}
+    context = {"page": page}
     return render(request, 'base/login_register.html', context)
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An Error occurred during regestraion.")
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 
 def logoutUser(request):
@@ -60,7 +84,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -72,10 +96,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are Not Allowed Here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -86,9 +113,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are Not Allowed Here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
